+++ packages/mailer/vite.config.ts
@@ -3,6 +3,8 @@ import { fileURLToPath } from "node:url";
 
 import { defineConfig, loadEnv } from "vite";
 
+import { dependencies, peerDependencies } from "./package.json";
+
 // https://vitejs.dev/config/
 export default defineConfig(({ mode }) => {
   process.env = { ...process.env, ...loadEnv(mode, process.cwd()) };
@@ -16,16 +18,8 @@ export default defineConfig(({ mode }) => {
       },
       rollupOptions: {
         external: [
-          "@dzangolab/fastify-config",
-          "fastify",
-          "fastify-plugin",
-          "html-minifier",
-          "html-to-text",
-          "mjml",
-          "mustache",
-          "nodemailer",
-          "nodemailer-html-to-text",
-          "nodemailer-mjml",
+          ...Object.keys(dependencies),
+          ...Object.keys(peerDependencies),
         ],
         output: {
           exports: "named",