+++ packages/slonik/src/types/index.ts
@@ -2,6 +2,6 @@ export type { SlonikConfig } from "./config";
 
 export type { Database, FilterInput, SortInput } from "./database";
 
-export type { Service } from "./service";
+export type { PaginatedList, Service } from "./service";
 
 export type { SqlFactory } from "./sqlFactory";