+++ packages/user/src/__test__/resolver.spec.ts
@@ -48,8 +48,15 @@ describe("user service resolver", () => {
 
       if (limit && offset) {
         resolver.Query.users(undefined, { limit, offset }, context);
-        /* eslint-disable-next-line unicorn/no-useless-undefined */
-        expect(list).toBeCalledWith(limit, offset, undefined, undefined);
+
+        expect(list).toBeCalledWith(
+          limit,
+          offset,
+          /* eslint-disable-next-line unicorn/no-useless-undefined */
+          undefined,
+          /* eslint-disable-next-line unicorn/no-useless-undefined */
+          undefined
+        );
       }
     }
   });