+++ packages/user/src/supertokens/recipes/config/third-party-email-password/emailPasswordSignIn.ts
@@ -1,33 +1,70 @@
+import UserRoles from "supertokens-node/recipe/userroles";
+
+import UserProfileService from "../../../../model/user-profiles/service";
 import Email from "../../../utils/email";
 
+import type {
+  User,
+  UserProfile,
+  UserProfileCreateInput,
+  UserProfileUpdateInput,
+} from "../../../../types";
 import type { FastifyInstance } from "fastify";
+import type { QueryResultRow } from "slonik";
 import type { RecipeInterface } from "supertokens-node/recipe/thirdpartyemailpassword";
 
 const emailPasswordSignIn = (
   originalImplementation: RecipeInterface,
   fastify: FastifyInstance
 ): RecipeInterface["emailPasswordSignIn"] => {
+  const { config, slonik } = fastify;
+
   return async (input) => {
     const originalEmail = input.email;
 
     input.email = Email.addTenantPrefix(
-      fastify.config,
+      config,
       input.email,
       input.userContext.tenant
     );
 
-    let originalResponse = await originalImplementation.emailPasswordSignIn(
+    const originalResponse = await originalImplementation.emailPasswordSignIn(
       input
     );
 
-    if (originalResponse.status === "OK") {
-      originalResponse = {
-        ...originalResponse,
-        user: { ...originalResponse.user, email: originalEmail },
-      };
+    if (originalResponse.status !== "OK") {
+      return originalResponse;
     }
 
-    return originalResponse;
+    const service: UserProfileService<
+      UserProfile & QueryResultRow,
+      UserProfileCreateInput,
+      UserProfileUpdateInput
+    > = new UserProfileService(config, slonik);
+
+    /* eslint-disable-next-line unicorn/no-null */
+    let profile: UserProfile | null = null;
+
+    try {
+      profile = await service.findById(originalResponse.user.id);
+    } catch {
+      // FIXME [OP 2022-AUG-22] Handle error properly
+      // DataIntegrityError
+    }
+
+    const { roles } = await UserRoles.getRolesForUser(originalResponse.user.id);
+
+    const user: User = {
+      ...originalResponse.user,
+      email: originalEmail,
+      profile,
+      roles,
+    };
+
+    return {
+      status: "OK",
+      user,
+    };
   };
 };