+++ packages/user/src/validator/__test__/password.spec.ts
@@ -1,17 +1,19 @@
-import { describe, expect, it } from "vitest";
+import { beforeEach, describe, expect, it } from "vitest";
 
 import validatePassword from "../password";
 
 import type { ApiConfig } from "@dzangolab/fastify-config";
 
 describe("validatePassword", () => {
-  const config = {
-    user: {
-      supertokens: {
-        validatorOptions: {},
+  let config = {} as unknown as ApiConfig;
+
+  beforeEach(() => {
+    config = {
+      user: {
+        password: {},
       },
-    },
-  } as unknown as ApiConfig;
+    } as unknown as ApiConfig;
+  });
 
   it("returns error object when password field is not present", async () => {
     const password = undefined as unknown as string;
@@ -30,7 +32,7 @@ describe("validatePassword", () => {
     const passwordValidation = validatePassword(password, config);
 
     expect(passwordValidation).toEqual({
-      message: "Passsword should contain minimum 8 characters",
+      message: "Password should contain minimum 8 characters",
       success: false,
     });
   });
@@ -41,13 +43,13 @@ describe("validatePassword", () => {
     const passwordValidation = validatePassword(password, config);
 
     expect(passwordValidation).toEqual({
-      message: "Passsword should contain minimum 8 characters",
+      message: "Password should contain minimum 8 characters",
       success: false,
     });
   });
 
   it("returns success object when password is strong", async () => {
-    const password = "aaaaaaaa";
+    const password = "aaaaaaa1";
 
     const passwordValidation = validatePassword(password, config);
 
@@ -57,32 +59,28 @@ describe("validatePassword", () => {
   });
 
   it("returns error object when password length less than minLength", async () => {
-    const strongPasswordOptions = {
+    config.user.password = {
       minLength: 10,
     };
 
-    config.user.supertokens.validatorOptions.password = strongPasswordOptions;
-
-    const password = "Qwerty";
+    const password = "Qwerty1";
 
     const passwordValidation = validatePassword(password, config);
 
     expect(passwordValidation).toEqual({
-      message: "Passsword should contain minimum 10 characters",
+      message: "Password should contain minimum 10 characters",
       success: false,
     });
   });
 
-  it("returns error objects when all provided options are not satsfied", async () => {
-    const strongPasswordOptions = {
+  it("returns error object when all provided options are not satsfied", async () => {
+    config.user.password = {
       minLength: 1,
       minLowercase: 1,
       minUppercase: 1,
       minNumbers: 1,
       minSymbols: 1,
     };
-
-    config.user.supertokens.validatorOptions.password = strongPasswordOptions;
 
     const password = "Qwerty12";
 
@@ -90,21 +88,19 @@ describe("validatePassword", () => {
 
     expect(passwordValidation).toEqual({
       message:
-        "Passsword should contain minimum 1 character, minimum 1 lowercase, minimum 1 uppercase, minimum 1 number and minimum 1 symbol",
+        "Password should contain minimum 1 character, minimum 1 lowercase, minimum 1 uppercase, minimum 1 number and minimum 1 symbol",
       success: false,
     });
   });
 
-  it("returns success object when all provided options are satsfied", async () => {
-    const strongPasswordOptions = {
+  it("returns success object when all provided options are satisfied", async () => {
+    config.user.password = {
       minLength: 1,
       minLowercase: 1,
       minUppercase: 1,
       minNumbers: 1,
       minSymbols: 1,
     };
-
-    config.user.supertokens.validatorOptions.password = strongPasswordOptions;
 
     const password = "Qwerty1!";
 
@@ -116,15 +112,13 @@ describe("validatePassword", () => {
   });
 
   it("returns error objects when custom strongPasswordOptions are provided", async () => {
-    const strongPasswordOptions = {
+    config.user.password = {
       minLength: 2,
       minLowercase: 2,
       minUppercase: 2,
       minNumbers: 2,
       minSymbols: 2,
     };
-
-    config.user.supertokens.validatorOptions.password = strongPasswordOptions;
 
     const password = "Qwerty12";
 
@@ -132,21 +126,19 @@ describe("validatePassword", () => {
 
     expect(passwordValidation).toEqual({
       message:
-        "Passsword should contain minimum 2 characters, minimum 2 lowercases, minimum 2 uppercases, minimum 2 numbers and minimum 2 symbols",
+        "Password should contain minimum 2 characters, minimum 2 lowercases, minimum 2 uppercases, minimum 2 numbers and minimum 2 symbols",
       success: false,
     });
   });
 
   it("returns success object when password pass the provided strongPasswordOptions", async () => {
-    const strongPasswordOptions = {
+    config.user.password = {
       minLength: 2,
       minLowercase: 2,
       minUppercase: 2,
       minNumbers: 2,
       minSymbols: 2,
     };
-
-    config.user.supertokens.validatorOptions.password = strongPasswordOptions;
 
     const password = "QwertY12!@";